Implement backup plugin & Remove command abstract class
