+++ app/server/skychat/Room.ts
@@ -1,9 +1,8 @@
 import {Connection} from "./Connection";
 import {IBroadcaster} from "./IBroadcaster";
 import {Message, MessageConstructorOptions, MessageMeta} from "./Message";
-import {Command} from "./commands/Command";
 import {Plugin} from "./commands/Plugin";
-import {CommandManager} from "./commands/CommandManager";
+import {PluginManager} from "./commands/PluginManager";
 import * as fs from "fs";
 import SQL from "sql-template-strings";
 import {DatabaseHelper} from "./DatabaseHelper";
@@ -65,21 +64,21 @@ export class Room implements IBroadcaster {
     public locked: boolean = false;
 
     /**
-     * Command instances (including plugins).
-     * All aliases of a command/plugin points to the same command instance.
+     * Plugins. All aliases of a command/plugin points to the same command instance.
      */
-    public readonly commands: {[commandName: string]: Command};
+    public readonly commands: {[commandName: string]: Plugin};
 
     /**
-     * List of loaded plugins
+     * Plugins sorted by descending priority
      */
     public readonly plugins: Plugin[];
 
     constructor(id: number, name: string) {
         this.id = id;
         this.name = name;
-        this.commands = CommandManager.instantiateCommands(this);
-        this.plugins = CommandManager.extractPlugins(this.commands);
+        const {commands, plugins} = PluginManager.instantiatePlugins(this);
+        this.commands = commands;
+        this.plugins = plugins;
         this.load();
     }
 
@@ -179,8 +178,8 @@ export class Room implements IBroadcaster {
      * Get a plugin instance by its name
      * @param name
      */
-    public getPlugin(name: string): Command {
-        const plugin = this.commands[name];
+    public getPlugin(name: string): Plugin {
+        const plugin: Plugin | undefined = this.commands[name];
         if (! plugin) {
             throw new Error('Plugin not found');
         }
