+++ package.json
@@ -18,7 +18,8 @@
         "kill": "ps aux | grep 'node --max-old-space-size=8192 build/server.js' | awk '{print $2}' | xargs kill -9",
         "dev": "npm run clean && ./node_modules/gulp/bin/gulp.js watch",
         "start": "npm run build && node build/server.js",
-        "start:docker": "bash scripts/docker-start.sh"
+        "start:docker": "bash scripts/docker-start.sh",
+        "backup": "bash scripts/backup.sh"
     },
     "keywords": [],
     "author": "",
